fix: MUST FIX項目の対応 - UX改善と共通化 (#63)
背景:
PR#70の3回目のレビュー（APPROVE with Recommendations）で
指摘されたMUST FIX項目を対応。

修正内容:

1. 配列の安全なアクセス修正 ✅
   - NamingTemplateSection.tsx: getTodayDate()
   - split('T')[0]からslice(0, 10)に変更
   - より明示的で型安全なアプローチ

2. Enterキーでの送信対応 ✅
   - AccountSection.tsx: ライセンスキー入力にEnterキー対応追加
   - handleKeyDown関数を実装
   - 基本的なUX要件を満たす

3. evaluateTemplate関数の共通化 ✅
   - src/shared/utils/template.ts を新規作成
   - テンプレート評価ロジックを共通化
   - NamingTemplateSection.tsxから重複コード削除
   - DRY原則準拠、将来のZIP生成時にも利用可能

技術的詳細:
- TypeScript型安全性の向上
- キーボード操作のアクセシビリティ改善
- コードの保守性向上（共通ユーティリティ関数）

テスト:
- ビルド: ✅ 成功
- TypeScript型チェック: ✅ 通過

関連Issue: #63
関連PR: #70
